+++ rotkehlchen/tasks/manager.py
@@ -48,7 +48,7 @@ from rotkehlchen.tasks.assets import (
 from rotkehlchen.tasks.calendar import (
     CalendarNotification,
     delete_past_calendar_entries,
-    maybe_create_ens_reminders,
+    maybe_create_calendar_reminders,
     notify_reminders,
 )
 from rotkehlchen.tasks.utils import query_missing_prices_of_base_entries, should_run_periodic_task
@@ -832,9 +832,9 @@ class TaskManager:
 
         return [self.greenlet_manager.spawn_and_track(
             after_seconds=None,
-            task_name='Maybe create ENS reminders',
+            task_name='Maybe create calendar reminders',
             exception_is_error=True,
-            method=maybe_create_ens_reminders,
+            method=maybe_create_calendar_reminders,
             database=self.database,
         )]